Refactored app.py to move event handling to events.py
